feat: Get rid of using class records for txs
