+++ src/js/records/index.ts
@@ -1,3 +1,2 @@
 export * from './models';
 export * from './monobank';
-export * from './transaction-model.record';
