+++ src/js/records/monobank/index.ts
@@ -1,3 +1,1 @@
 export { MONOUserRecord } from './user.record';
-export { MONOUserCurrencyRecord } from './currency.record';
-export { MONOTransactionRecord } from './transaction.record';
